fw/apps/prf: remove extras menu, move Load PRF to main menu
Remove the extras menu entirely and move its only remaining item
(Load PRF) to the main menu as the last option.
